move to cmake build
