+++ .gitignore
@@ -1,2 +1,3 @@
 /www/
 /node_modules/
+/build/
