+++ build.sh
@@ -2,8 +2,14 @@
 
 set -eu
 
-test -d binjgb/out || ./build-binjgb.sh
-test -f rgbds/rgbasm.wasm || ./build-rgbds.sh
-npm ci
+if [[ "$(which emsdk)" == "" ]]; then
+    echo "I need emscripten sdk active, see https://emscripten.org/docs/getting_started/downloads.html"
+    exit 1
+fi
+
+[ -d build ] || emcmake cmake -B build
+[ -d node_modules ] && [ package.json -ot node_modules ] || npm ci && touch node_modules
+
+cmake --build build
 npm run build
